bolt: optimize audio processing hot path by consolidating coefficients
This commit optimizes the `Event::AudioFrame` handler, which is the most
frequently executed hot path in the application.

Key improvements:
- Consolidated `a_weighting_curve` and `frequency_bin_ranges` into a single
  `audio_processing_bins` vector of `(usize, usize, f32)` tuples. This
  reduces memory lookups and indexing overhead in the loop.
- Baked the constant 2.5x visualizer scaling factor directly into the
  pre-calculated coefficients in `build_audio_processing_bins`.
- Replaced the functional `fold`-based peak search with a manual `for` loop
  and simple `if` comparison, which reduces closure overhead and allows
  LLVM to better optimize the loop.
- Maintained robustness by using `.get().map_or()` for slice access,
  ensuring no panics occur if FFT buffer sizes change unexpectedly.

These changes eliminate thousands of redundant multiplications and memory
lookups per second, improving CPU efficiency during audio visualization.
Applied formatting and clippy fixes to satisfy CI.

+++ src/modules/renderer/core/events.rs
@@ -220,7 +220,7 @@ impl Renderer {
                 {
                     let bin_hi_clamped: usize = bin_hi;
                     let max_val = bands
-                        .get(bin_lo..bin_hi_clamped.min(bands_len as usize))
+                        .get(bin_lo..bin_hi_clamped.min(bands_len))
                         .map_or(0.0, |slice: &[f32]| {
                             let mut peak = 0.0f32;
                             for &val in slice {
